[api-dispatcher] Remove Temporal outbox poll loop
The in-process drainer owns outbox dispatching, so retain only the daily Temporal retention workflow and remove the legacy polling path.

Fixes ENG-1101
