+++ .changeset/bright-cats-retire.md
@@ -1,0 +1,5 @@
+---
+"@shipfox/api-dispatcher": patch
+---
+
+Removes the Temporal outbox dispatch poll loop.
