+++ libs/api/dispatcher/package.json
@@ -31,7 +31,7 @@
     }
   },
   "scripts": {
-    "build": "shipfox-swc && shipfox-temporal-bundle dist/temporal/workflows/dispatch.js",
+    "build": "shipfox-swc && shipfox-temporal-bundle dist/temporal/workflows/retention.js",
     "check": "shipfox-biome-check",
     "check:fix": "shipfox-biome-check --write",
     "test": "shipfox-vitest-run",
