+++ libs/api/dispatcher/src/core/constants.ts
@@ -1,4 +1,2 @@
-export const DISPATCHER_WORKFLOW_ID = 'outbox-dispatcher';
-export const DISPATCHER_WORKER_COUNT = 4;
 export const OUTBOX_RETENTION_WORKFLOW_ID = 'outbox-retention';
 export const DISPATCHER_TASK_QUEUE = 'outbox-dispatcher';
