+++ libs/api/dispatcher/src/core/index.ts
@@ -1,5 +1,1 @@
-export {
-  DISPATCHER_TASK_QUEUE,
-  DISPATCHER_WORKER_COUNT,
-  DISPATCHER_WORKFLOW_ID,
-} from './constants.js';
+export {DISPATCHER_TASK_QUEUE} from './constants.js';
