+++ libs/api/dispatcher/src/index.ts
@@ -1,3 +1,3 @@
-export {DISPATCHER_TASK_QUEUE, DISPATCHER_WORKFLOW_ID} from '#core/index.js';
+export {DISPATCHER_TASK_QUEUE} from '#core/index.js';
 export {createActivities} from '#temporal/index.js';
 export {dispatcherModule} from './module.js';
