--- libs/api/dispatcher/src/temporal/activities/drain-and-dispatch.ts
@@ -1,1 +1,0 @@
-export {runDrainCycle as drainAndDispatch} from '#core/run-drain-cycle.js';
