+++ libs/api/dispatcher/src/temporal/activities/index.ts
@@ -1,9 +1,7 @@
-import {drainAndDispatch} from './drain-and-dispatch.js';
 import {pruneOutboxRetention} from './prune-outbox-retention.js';
 
 export function createActivities() {
   return {
-    drainAndDispatch,
     pruneOutboxRetention,
   };
 }
